feat/#93: 일지 상세조회 api 개발

+++ src/main/java/com/epi/epilog/app/domain/log/Log.java
@@ -65,7 +65,7 @@ public class Log extends BaseEntity {
     @Nullable
     private Double diastolicBloodPressure; // 이완기 혈압
     @Nullable
-    private Integer heartRate; // 심박수
+    private Double heartRate; // 심박수
 
     @Nullable
     private Double weight; // 몸무게
